MDEXP-473 - CQL export fails when effective location is included in the cql statement
