+++ src/main/java/org/folio/clients/InventoryClient.java
@@ -6,34 +6,22 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import javax.ws.rs.core.MediaType;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.http.HttpHeaders;
-import org.apache.http.HttpStatus;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.folio.service.logs.ErrorLogService;
 import org.folio.util.ErrorCode;
 import org.folio.util.OkapiConnectionParams;
-import org.folio.util.StringUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import io.vertx.core.Future;
-import io.vertx.core.Promise;
-import io.vertx.core.buffer.Buffer;
-import io.vertx.core.json.DecodeException;
 import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
-import io.vertx.ext.web.client.HttpRequest;
-import io.vertx.ext.web.client.HttpResponse;
 import io.vertx.ext.web.client.WebClient;
 
 import static java.lang.String.format;
 import static org.folio.clients.ClientUtil.buildQueryEndpoint;
 import static org.folio.clients.ClientUtil.getRequest;
-import static org.folio.rest.RestVerticle.OKAPI_HEADER_TENANT;
-import static org.folio.rest.RestVerticle.OKAPI_HEADER_TOKEN;
 import static org.folio.util.ExternalPathResolver.ALTERNATIVE_TITLE_TYPES;
 import static org.folio.util.ExternalPathResolver.CALL_NUMBER_TYPES;
 import static org.folio.util.ExternalPathResolver.CAMPUSES;
@@ -54,7 +42,6 @@ import static org.folio.util.ExternalPathResolver.LIBRARIES;
 import static org.folio.util.ExternalPathResolver.LOAN_TYPES;
 import static org.folio.util.ExternalPathResolver.LOCATIONS;
 import static org.folio.util.ExternalPathResolver.MATERIAL_TYPES;
-import static org.folio.util.ExternalPathResolver.RECORD_BULK_IDS;
 import static org.folio.util.ExternalPathResolver.resourcesPathWithPrefix;
 
 @Component
@@ -65,11 +52,6 @@ public class InventoryClient {
   private static final String QUERY_LIMIT_PATTERN = "?query=(%s)&limit=";
   private static final String QUERY_PATTERN_HOLDING = "instanceId==%s";
   private static final String QUERY_PATTERN_ITEM = "holdingsRecordId==%s";
-  private static final String QUERY = "?query=";
-  private static final String ERROR_MESSAGE_INVALID_STATUS_CODE = "Exception while calling %s, message: Get invalid response with status: %s";
-  private static final String ERROR_MESSAGE_INVALID_BODY = "Exception while calling %s, message: Got invalid response body: %s";
-  private static final String ERROR_MESSAGE_EMPTY_BODY = "Exception while calling %s, message: empty body returned.";
-  private static final String BULK_EDIT_HOLDING_QUERY_PREFIX = "?field=instanceId&recordType=HOLDING&imit=10&query=";
   private static final int REFERENCE_DATA_LIMIT = 1000;
   private static final int HOLDINGS_LIMIT = 1000;
 
@@ -87,55 +69,6 @@ public class InventoryClient {
       errorLogService.saveGeneralErrorWithMessageValues(ErrorCode.ERROR_GETTING_INSTANCES_BY_IDS.getCode(), Arrays.asList(exception.getMessage()), jobExecutionId, params.getTenantId());
       return Optional.empty();
     }
-  }
-
-  public Future<Optional<JsonObject>> getInstancesBulkUUIDsAsync(String query, OkapiConnectionParams params) {
-    Promise<Optional<JsonObject>> promise = Promise.promise();
-    if (StringUtils.isEmpty(query)) {
-      promise.complete(Optional.empty());
-      return promise.future();
-    }
-    String endpoint = format(resourcesPathWithPrefix(RECORD_BULK_IDS), params.getOkapiUrl()) + QUERY + StringUtil.urlEncode(query);
-    HttpRequest<Buffer> request = webClient.getAbs(endpoint);
-    request.putHeader(OKAPI_HEADER_TOKEN, params.getToken());
-    request.putHeader(OKAPI_HEADER_TENANT, params.getTenantId());
-    request.putHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON);
-    request.putHeader(HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON);
-    if (params.getOkapiUrl().contains("https")) {
-      request.ssl(true);
-    }
-    request.send(res -> {
-      if (res.failed()) {
-        logError(res.cause(), params);
-        promise.complete(Optional.empty());
-      } else {
-        HttpResponse<Buffer> response = res.result();
-        if (response.statusCode() != HttpStatus.SC_OK) {
-          logError(new IllegalStateException(format(ERROR_MESSAGE_INVALID_STATUS_CODE, endpoint, response.statusCode())), params);
-          promise.complete(Optional.empty());
-        } else {
-          try {
-            JsonObject instances = response.bodyAsJsonObject();
-            if (instances != null) {
-              promise.complete(Optional.of(instances));
-            } else {
-              logError(new IllegalStateException(format(ERROR_MESSAGE_EMPTY_BODY, endpoint)), params);
-              promise.complete(Optional.empty());
-            }
-          } catch (DecodeException ex) {
-            LOGGER.debug("Cannot process instances, invalid json body returned.", ex);
-            logError(new IllegalStateException(format(ERROR_MESSAGE_INVALID_BODY, endpoint, response.bodyAsString())), params);
-            promise.complete(Optional.empty());
-          }
-        }
-      }
-    });
-    return promise.future();
-  }
-
-  private void logError(Throwable throwable, OkapiConnectionParams params) {
-    LOGGER.error(throwable.getMessage(), throwable.getCause());
-    errorLogService.saveGeneralErrorWithMessageValues(ErrorCode.ERROR_GETTING_INSTANCES_BY_IDS.getCode(), Arrays.asList(throwable.getMessage()), StringUtils.EMPTY, params.getTenantId());
   }
 
   public Map<String, JsonObject> getNatureOfContentTerms(String jobExecutionId, OkapiConnectionParams params) {
